Split Main.hs into smaller pieces
